move removeNullValues to shared package
