+++ apps/web/src/components/GalleryEditor/getInitialCollectionsFromServer.ts
@@ -7,9 +7,9 @@ import {
   StagedSectionMap,
 } from '~/components/GalleryEditor/GalleryEditorContext';
 import { getInitialCollectionsFromServerFragment$key } from '~/generated/getInitialCollectionsFromServerFragment.graphql';
+import { removeNullValues } from '~/shared/relay/removeNullValues';
 import { parseCollectionLayoutGraphql } from '~/utils/collectionLayout';
 import { generate12DigitId } from '~/utils/generate12DigitId';
-import { removeNullValues } from '~/utils/removeNullValues';
 
 export function getInitialCollectionsFromServer(
   galleryRef: getInitialCollectionsFromServerFragment$key
