+++ apps/web/src/contexts/globalLayout/GlobalAnnouncementPopover/components/GalleryOfTheWeekCard.tsx
@@ -12,8 +12,8 @@ import FollowButton from '~/components/Follow/FollowButton';
 import { GalleryOfTheWeekCardQueryFragment$key } from '~/generated/GalleryOfTheWeekCardQueryFragment.graphql';
 import { GalleryOfTheWeekCardUserFragment$key } from '~/generated/GalleryOfTheWeekCardUserFragment.graphql';
 import { useIsMobileOrMobileLargeWindowWidth } from '~/hooks/useWindowSize';
+import { removeNullValues } from '~/shared/relay/removeNullValues';
 import getVideoOrImageUrlForNftPreview from '~/utils/graphql/getVideoOrImageUrlForNftPreview';
-import { removeNullValues } from '~/utils/removeNullValues';
 
 type GalleryOfTheWeekCardProps = {
   queryRef: GalleryOfTheWeekCardQueryFragment$key;
