+++ apps/web/src/scenes/MerchStorePage/Redemption/RedeemModal.tsx
@@ -8,7 +8,7 @@ import { HStack, VStack } from '~/components/core/Spacer/Stack';
 import { TitleXS } from '~/components/core/Text/Text';
 import { MerchType } from '~/generated/getObjectNameFragment.graphql';
 import { RedeemModalQueryFragment$key } from '~/generated/RedeemModalQueryFragment.graphql';
-import { removeNullValues } from '~/utils/removeNullValues';
+import { removeNullValues } from '~/shared/relay/removeNullValues';
 
 import RedeemedPage from './RedeemedPage';
 import ToRedeemPage from './ToRedeemPage';
